Refactor subscription components: add new profile creation and balance check features, enhance payment method handling, and improve state management.
