+++ deploily-console/src/app/[locale]/(auth)/portal/api-services/[id]/components/subscriptionDrawer/containers/paymentComponent.tsx
@@ -3,34 +3,35 @@ import { Flex, Radio, RadioChangeEvent, Typography, } from "antd";
 import { theme } from "@/styles/theme";
 import BankTransfertComponent from "../components/bankTransfertComponent";
 import { useScopedI18n } from "../../../../../../../../../../locales/client";
-import { useAppDispatch } from "@/lib/hook";
 import CardPaymentComponent from "../components/cardPaymentComponent";
-import { useSubscriptionStates } from "@/lib/features/subscription-states/subscriptionSelectors";
+import { useState } from "react";
 
 export default function PaymentComponent({ selectedPlan }: { selectedPlan: any }) {
-  const dispatch = useAppDispatch();
-  const { payment_method } = useSubscriptionStates()
-
+  const translate = useScopedI18n('subscription');
+  const [paymentMethod, setPaymentMethod] = useState("card")
   const onChange = (e: RadioChangeEvent) => {
-    dispatch({ type: "SubscriptionStates/updateSubscriptionStates", payload: { "payment_method": e.target.value } })
+    setPaymentMethod(e.target.value );
   };
-
-
   const t = useScopedI18n("payments");
-
+  
   return (
     <>
+      <Typography.Text style={{
+        color: theme.token.red500, paddingTop: 30, display: "flex",
+        justifyContent: "center",
+      }}>{translate("insufficientBalance")}
+      </Typography.Text>
       <Typography.Title level={4} style={{ paddingTop: 20, paddingBottom: 20 }}>Choose the payment method</Typography.Title>
       <Flex vertical gap="start" style={{ padding: 10, backgroundColor: theme.token.colorBgBase, }}>
-        <Radio.Group block defaultValue={payment_method}
+        <Radio.Group block defaultValue={paymentMethod}
           onChange={onChange}
-          value={payment_method}>
+          value={paymentMethod}>
           <Radio value="card"  >{t("card")}</Radio>
           <Radio value="bank_transfer">{t("bank")}</Radio>
         </Radio.Group>
 
       </Flex>
-      {payment_method === "card" ?
+      {paymentMethod === "card" ?
         <CardPaymentComponent selectedPlan={selectedPlan}/>
         : <BankTransfertComponent selectedPlan={selectedPlan} />}
     </>
